+++ src/main/java/pl/dofizjo/dofizjoapplication/data/mapper/ReviewMapper.java
@@ -11,7 +11,7 @@ public class ReviewMapper implements RowMapper<Review> {
     @Override
     public Review mapRow(ResultSet rs, int rowNum) throws SQLException {
         Review review = new Review();
-        review.setId(rs.getLong("id"));
+        review.setId(rs.getInt("id"));
         review.setAuthor(rs.getString("author"));
         review.setDiscipline(rs.getString("discipline"));
         review.setOpinion(rs.getString("opinion"));
